pyproject.toml
[project]
name = "fiap-hackaton-5"
version = "0.1.0"
description = ""
authors = [
    {name = "Leandro Bernardo",email = "leandro.bernardos@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.54.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "openai (>=2.21.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
